Get hostname #506 Rename file
