+++ src/hooks.server.ts
@@ -4,7 +4,7 @@ import { Repository } from '$lib/app/repository'
 import { Session } from '$lib/auth/session'
 import { Signing } from '$lib/auth/signing'
 import { ClientAddress } from '$lib/network/client_address'
-import { ClientHostName } from '$lib/network/client_hostnames'
+import { ClientHostName } from '$lib/network/client_hostname'
 import type { AuthToken, Role, User } from '@prisma/client'
 import type { Handle, HandleServerError, RequestEvent } from '@sveltejs/kit'
